Refactor Opp bookmark into separate component
