+++ components/OppBookmark/styles.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+let styles = StyleSheet.create({
+  oppBookmark: {
+    position: 'absolute',
+    top: 0,
+    right: 2
+  }
+});
+
+export default styles;
